Move URL credential helpers to a leaf module, break import cycle
Teaching createConnection to decode URL components made zeroSetup.ts import
decodeUrlComponent from bootstrap.ts, while bootstrap.ts already imported
channelFromURL from zeroSetup.ts -- a cycle that the Rollup CLI build reported as
a new circular-dependency warning. The cycle was runtime-safe (both symbols are
referenced only inside function bodies) but the warning is avoidable. Relocate
redactUrlCredentials and decodeUrlComponent, which depend on nothing in either
command module, into a new leaf util apps/cli/src/util/connectionUrl.ts that both
commands and invite.ts import from. bootstrap.ts keeps its one-directional import
of channelFromURL, so no cycle remains and the build is warning-free again.

Also extend the unit coverage the helper move touches: the reconcile-agreement
test now uses a percent-encoded host so a decoded host actually flows into
diffConnectionAgainstTarget, and a new test pins that an encoded slash (%2F) in
an SFTP path decodes to a separator -- the intended literal, since a POSIX
filename cannot contain a slash.
